tbd-detector/detector_requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
tensorflow-cpu==2.17.0
numpy>=1.26.0,<2.0.0
opencv-python-headless==4.8.1.78
python-multipart